news: split comment edit tasks
