+++ handlers/news/newsCommentEditTasks.go
@@ -1,109 +1,4 @@
 package news
 
-import (
-	"database/sql"
-	"fmt"
-	"github.com/arran4/goa4web/core/consts"
-	"net/http"
-	"strconv"
-
-	"github.com/gorilla/mux"
-
-	"github.com/arran4/goa4web/core"
-	"github.com/arran4/goa4web/core/common"
-	"github.com/arran4/goa4web/handlers"
-	"github.com/arran4/goa4web/internal/db"
-	notif "github.com/arran4/goa4web/internal/notifications"
-	"github.com/arran4/goa4web/internal/tasks"
-	"github.com/arran4/goa4web/workers/postcountworker"
-)
-
-// EditReplyTask updates an existing comment.
-type EditReplyTask struct{ tasks.TaskString }
-
-var editReplyTask = &EditReplyTask{TaskString: TaskEditReply}
-
-var _ tasks.Task = (*EditReplyTask)(nil)
-var _ notif.AdminEmailTemplateProvider = (*EditReplyTask)(nil)
-
-func (EditReplyTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationNewsCommentEditEmail")
-}
-
-func (EditReplyTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationNewsCommentEditEmail")
-	return &v
-}
-
-func (EditReplyTask) Action(w http.ResponseWriter, r *http.Request) any {
-	languageId, err := strconv.Atoi(r.PostFormValue("language"))
-	if err != nil {
-		return fmt.Errorf("languageId parse fail %w", handlers.ErrRedirectOnSamePageHandler(err))
-	}
-	text := r.PostFormValue("replytext")
-
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	vars := mux.Vars(r)
-	postId, _ := strconv.Atoi(vars["post"])
-	commentId, _ := strconv.Atoi(vars["comment"])
-
-	session, ok := core.GetSessionOrFail(w, r)
-	if !ok {
-		return handlers.SessionFetchFail{}
-	}
-	uid, _ := session.Values["UID"].(int32)
-
-	comment := r.Context().Value(consts.KeyComment).(*db.GetCommentByIdForUserRow)
-
-	thread, err := queries.GetThreadLastPosterAndPerms(r.Context(), db.GetThreadLastPosterAndPermsParams{
-		ViewerID:      uid,
-		ThreadID:      comment.ForumthreadID,
-		ViewerMatchID: sql.NullInt32{Int32: uid, Valid: uid != 0},
-	})
-	if err != nil {
-		return fmt.Errorf("thread fetch fail %w", handlers.ErrRedirectOnSamePageHandler(err))
-	}
-
-	if err = queries.UpdateComment(r.Context(), db.UpdateCommentParams{
-		Idcomments:         int32(commentId),
-		LanguageIdlanguage: int32(languageId),
-		Text:               sql.NullString{String: text, Valid: true},
-	}); err != nil {
-		return fmt.Errorf("update comment fail %w", handlers.ErrRedirectOnSamePageHandler(err))
-	}
-
-	if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-		if evt := cd.Event(); evt != nil {
-			if evt.Data == nil {
-				evt.Data = map[string]any{}
-			}
-			evt.Data[postcountworker.EventKey] = postcountworker.UpdateEventData{ThreadID: thread.Idforumthread, TopicID: thread.ForumtopicIdforumtopic}
-			evt.Data["CommentURL"] = cd.AbsoluteURL(fmt.Sprintf("/news/news/%d", postId))
-		}
-	}
-
-	return handlers.RedirectHandler(fmt.Sprintf("/news/news/%d", postId))
-}
-
-// CancelTask cancels comment editing.
-type CancelTask struct{ tasks.TaskString }
-
-var cancelTask = &CancelTask{TaskString: TaskCancel}
-
-var _ tasks.Task = (*CancelTask)(nil)
-var _ notif.AdminEmailTemplateProvider = (*CancelTask)(nil)
-
-func (CancelTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationNewsCommentCancelEmail")
-}
-
-func (CancelTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationNewsCommentCancelEmail")
-	return &v
-}
-
-func (CancelTask) Action(w http.ResponseWriter, r *http.Request) any {
-	vars := mux.Vars(r)
-	postId, _ := strconv.Atoi(vars["post"])
-	return handlers.RedirectHandler(fmt.Sprintf("/news/news/%d", postId))
-}
+// Helper functions shared by comment edit tasks live in this file. At present
+// no helpers are required.
